pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cointrader-trainer"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "numpy",
    "pyarrow",
    "pytz",
    "tenacity",
    "requests",
]

[project.optional-dependencies]
train = ["lightgbm", "scikit-learn", "optuna", "networkx", "backtrader"]
gpu = ["pyopencl"]
flwr = ["flwr"]
talib = ["TA-Lib"]
accel = ["numba", "jax"]
dev = ["ruff", "vulture", "pytest-cov"]

[project.scripts]
cointrainer = "cointrainer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
